Replace content reaction and form dialog with content feedback
